# docs/openapi.yaml
...
components:
...
    schemas:
...
        ErrorBody:
            properties:
                code:
                    type: string
                field:
                    type: string
                message:
                    type: string
                request_id:
                    type: string
            type: object
        ErrorEnvelope:
            properties:
                error:
                    $ref: '#/components/schemas/ErrorBody'
            type: object
        Favorite:
            properties:
                added_at:
                    type: string
                item_id:
                    type: string
                owner_id:
                    type: string
                photo_url:
                    type: string
                price:
                    type: integer
                status:
                    type: string
                title:
                    type: string
            type: object
        FavoriteListResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/Favorite'
                    type: array
                    uniqueItems: false
                next_cursor:
                    type: string
            type: object
...
    securitySchemes:
        bearerAuth:
            bearerFormat: JWT
            description: |
                JWT, подписанный HS256. Клеймы: `sub` — UUID пользователя,
                `role` — роль, `exp` — момент истечения, `iat` — момент выпуска,
                `jti` — уникальный идентификатор токена.

                Просроченный токен и токен с неверной подписью одинаково
                дают 401 с кодом `unauthorized`.
            in: header
            name: Authorization
            scheme: bearer
            type: http
info:
    contact:
        name: Avito Hack Team
    description: |
        API объявлений с игровым слоем: пользователь заводит объявления, добавляет фотографии,
        ведёт избранное, а параллельно развивает питомца — за действия в маркетплейсе
        начисляется опыт, растут уровень и стрик, открываются награды.

        ## Конверт ошибок

        Все ошибки возвращаются единым конвертом `ErrorEnvelope`:

        ```json
        {"error": {"code": "validation_error", "message": "title must be at least 3 characters", "field": "title", "request_id": "..."}}
        ```

        Поля `field` и `request_id` опциональны и опускаются, когда пусты.
        Коды: `validation_error`, `bad_request`, `unauthorized`, `forbidden`,
        `not_found`, `conflict`, `internal_error`.

        ## Авторизация

        JWT (HS256) передаётся в заголовке `Authorization: Bearer <token>`.
        Срок жизни токена лежит в клейме `exp` внутри самого токена — отдельного поля
        `expires_at` в ответах на регистрацию и вход нет. В клеймах также присутствуют
        `sub` (идентификатор пользователя) и `role` (`user` | `moderator` | `admin`).
        WebSocket авторизуется query-параметром `token`, так как браузерный
        WebSocket API не позволяет задавать заголовки.

        ## Пагинация

        Списки используют keyset-курсоры, непрозрачные для клиента (base64url).
        В проекте два независимых формата курсора:

...

...

        ## Деньги

...

        ## Тело запроса

...
    license:
...
openapi: 3.1.0
paths:
...
    /api/v1/favorites:
        get:
            description: |-
                Избранные объявления владельца токена, свежие сверху.
                Пагинация — временной курсор по времени добавления в избранное.
                Поле `photo_url` содержит первую фотографию объявления и опускается,
                если фотографий нет.
            operationId: listFavorites
            parameters:
                - description: Размер страницы. По умолчанию 20, максимум 100.
                  in: query
                  name: limit
                  schema:
                    default: 20
                    maximum: 100
                    type: integer
                - description: Временной курсор из `next_cursor` предыдущей страницы.
                  in: query
                  name: cursor
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FavoriteListResponse'
                    description: Страница избранного
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Список избранного
            tags:
                - Favorites
...
    /api/v1/items/{id}/favorite:
        delete:
            description: |-
                Идемпотентна: удаление отсутствующей записи также даёт 204.
                Тело ответа пустое. Опыт за удаление не начисляется и не отнимается.
            operationId: removeFavorite
            parameters:
                - description: Идентификатор объявления
                  in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "204":
                    description: Убрано из избранного
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Убрать из избранного
            tags:
                - Favorites
        post:
            description: |-
                Идемпотентна: повторное добавление того же объявления не создаёт дубль
                и возвращает те же 204 (ограничение уникальности пары пользователь-объявление
                обрабатывается как no-op). Тело ответа пустое.

                Побочный эффект: первое добавление порождает доменное событие,
                за которое питомцу асинхронно начисляется опыт. Повторное добавление
                опыт не приносит.
            operationId: addFavorite
            parameters:
                - description: Идентификатор объявления
                  in: path
                  name: id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "204":
                    description: Добавлено в избранное
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Добавить в избранное
            tags:
                - Favorites
...
servers:
...